Move WorkerFlags to WorkerSystem Remove ViewStorage types
