--- workers/unity/Packages/io.improbable.gdk.core/View/View.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 66956872847b3a34eb3136d138991431
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
