--- workers/unity/Packages/io.improbable.gdk.core/View/ViewStorage.cs
@@ -1,27 +1,0 @@
-using System;
-using Improbable.Worker.CInterop;
-
-namespace Improbable.Gdk.Core
-{
-    public interface IViewStorage
-    {
-        Type GetSnapshotType();
-        Type GetUpdateType();
-        uint GetComponentId();
-
-        bool HasComponent(long entityId);
-        Authority GetAuthority(long entityId);
-
-        void ApplyDiff(ViewDiff viewDiff);
-    }
-
-    public interface IViewComponentStorage<T> where T : struct, ISpatialComponentSnapshot
-    {
-        T GetComponent(long entityId);
-    }
-
-    public interface IViewComponentUpdater<T> where T : struct, ISpatialComponentUpdate
-    {
-        void ApplyUpdate(long entityId, in T update);
-    }
-}
