--- workers/unity/Packages/io.improbable.gdk.core/View/ViewStorage.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4ee4ba1a5f9a19446a900b72fd168c4d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
